Support clipMedia being renamed to livePostClipMedia
